# docs/swagger.yaml
definitions:
  models.Todos:
    properties:
      completed:
        type: boolean
      id:
        type: integer
      title:
        type: string
    type: object
info:
  contact: {}
paths:
  /todos:
    get:
      description: Get a list of todos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Todos'
            type: array
      summary: List todos
    post:
      consumes:
      - application/json
      description: Create a new todo
      parameters:
      - description: Todo input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Todos'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Todos'
      summary: Create a todo
  /todos/{id}:
    delete:
      description: Delete a todo by ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Todos'
      summary: Delete a todo by ID
    get:
      description: Get a todo by ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Todos'
      summary: Get a todo by ID
    put:
      consumes:
      - application/json
      description: Update a todo by ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      - description: Todo input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Todos'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Todos'
      summary: Update a todo by ID
swagger: "2.0"